Added contextswitch function, small bug in SRT and fixed it.
I added the contextSwitch Function function.
I added an executeProcess Function.
fixed a bug in SRT where it ran infinitely and it was because of a logical conditioning error where it was checking for process time to equal 0 instead of less than or equal to 0. Also, Removed the execute's function mutation part and moved that into the context switching function.